# src/main/java/dev/velocirawesome/etl/model/dto/ErrorResponse.java
package dev.velocirawesome.etl.model.dto;

public class ErrorResponse {
    private String error;
    private String message;

    public ErrorResponse() {
    }

    public ErrorResponse(String error, String message) {
        this.error = error;
        this.message = message;
    }

    public String getError() {
        return error;
    }

    public void setError(String error) {
        this.error = error;
    }

    public String getMessage() {
        return message;
    }

    public void setMessage(String message) {
        this.message = message;
    }
}